- move db/index to redux/save-state
