+++ packages/gatsby/src/state-machines/develop/actions.ts
@@ -10,7 +10,7 @@ import { IBuildContext } from "../../services"
 import { actions } from "../../redux/actions"
 import { listenForMutations } from "../../services/listen-for-mutations"
 import { DataLayerResult } from "../data-layer"
-import { saveState } from "../../db"
+import { saveState } from "../../redux/save-state"
 import reporter from "gatsby-cli/lib/reporter"
 import { store } from "../../redux"
 import { ProgramStatus } from "../../redux/types"
